requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
appdirs==1.4.4
    # via virtualenv
argparse==1.4.0
    # via unittest2
astroid==2.4.2
    # via
    #   pylint
    #   pylint-celery
attrs==20.3.0
    # via pytest
certifi==2020.12.5
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.14.4
    # via
    #   -r requirements/base.txt
    #   cryptography
chardet==4.0.0
    # via
    #   -r requirements/base.txt
    #   requests
click-log==0.3.2
    # via edx-lint
click==7.1.2
    # via
    #   click-log
    #   edx-lint
coverage==5.3.1
    # via
    #   -r requirements/test.in
    #   pytest-cov
cryptography==3.3.1
    # via
    #   -r requirements/base.txt
    #   social-auth-core
defusedxml==0.6.0
    # via
    #   -r requirements/base.txt
    #   python3-openid
    #   social-auth-core
distlib==0.3.1
    # via virtualenv
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
edx-lint==1.6
    # via -r requirements/test.in
filelock==3.0.12
    # via
    #   tox
    #   virtualenv
future==0.18.2
    # via pyjwkest
httpretty==1.0.5
    # via -r requirements/test.in
idna==2.10
    # via
    #   -r requirements/base.txt
    #   requests
iniconfig==1.1.1
    # via pytest
isort==5.7.0
    # via pylint
lazy-object-proxy==1.4.3
    # via astroid
linecache2==1.0.0
    # via traceback2
mccabe==0.6.1
    # via pylint
oauthlib==3.1.0
    # via
    #   -r requirements/base.txt
    #   requests-oauthlib
    #   social-auth-core
packaging==20.8
    # via
    #   pytest
    #   tox
pluggy==0.13.1
    # via
    #   pytest
    #   tox
py==1.10.0
    # via
    #   pytest
    #   tox
pycodestyle==2.6.0
    # via -r requirements/test.in
pycparser==2.20
    # via
    #   -r requirements/base.txt
    #   cffi
pycryptodomex==3.9.9
    # via pyjwkest
pyjwkest==1.4.2
    # via -r requirements/test.in
pyjwt==1.7.1
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
    #   social-auth-core
pylint-celery==0.3
    # via edx-lint
pylint-django==2.3.0
    # via edx-lint
pylint-plugin-utils==0.6
    # via
    #   pylint-celery
    #   pylint-django
pylint==2.6.0
    # via
    #   edx-lint
    #   pylint-celery
    #   pylint-django
    #   pylint-plugin-utils
pyparsing==2.4.7
    # via packaging
pytest-cov==2.10.1
    # via -r requirements/test.in
pytest-django==4.1.0
    # via -r requirements/test.in
pytest==6.2.1
    # via
    #   pytest-cov
    #   pytest-django
python3-openid==3.2.0
    # via
    #   -r requirements/base.txt
    #   social-auth-core
pytz==2020.5
    # via
    #   -r requirements/base.txt
    #   django
requests-oauthlib==1.3.0
    # via
    #   -r requirements/base.txt
    #   social-auth-core
requests==2.25.1
    # via
    #   -r requirements/base.txt
    #   pyjwkest
    #   requests-oauthlib
    #   social-auth-core
six==1.15.0
    # via
    #   -r requirements/base.txt
    #   astroid
    #   cryptography
    #   edx-lint
    #   pyjwkest
    #   social-auth-app-django
    #   social-auth-core
    #   tox
    #   unittest2
    #   virtualenv
social-auth-app-django==4.0.0
    # via -r requirements/base.txt
social-auth-core==4.0.2
    # via
    #   -r requirements/base.txt
    #   social-auth-app-django
sqlparse==0.4.1
    # via
    #   -r requirements/base.txt
    #   django
toml==0.10.2
    # via
    #   pylint
    #   pytest
    #   tox
tox==3.21.0
    # via -r requirements/test.in
traceback2==1.4.0
    # via unittest2
unittest2==1.1.0
    # via -r requirements/test.in
urllib3==1.26.2
    # via
    #   -r requirements/base.txt
    #   requests
virtualenv==20.3.0
    # via tox
wrapt==1.12.1
    # via astroid
